feat: main issue problem was not detected, nevertheless code refactoring was done
