+++ Streetcode/Streetcode.WebApi/Extensions/ConfigureHostBuilderExtensions.cs
@@ -1,13 +1,11 @@
 ﻿using System.IO.Compression;
-using System.Runtime;
-using Microsoft.Extensions.Configuration;
 using Serilog;
-using Streetcode.BLL.Middleware;
 using AspNetCoreRateLimit;
 using Microsoft.AspNetCore.ResponseCompression;
 using Streetcode.BLL.Services.BlobStorageService;
 using Streetcode.BLL.Services.Instagram;
 using Streetcode.BLL.Services.Payment;
+using Streetcode.WebApi.Middleware.ApiRequestResponseMiddleware;
 
 namespace Streetcode.WebApi.Extensions;
