+++ Streetcode/Streetcode.WebApi/Middleware/ApiRequestResponseMiddleware/ApiRequestResponseMiddleware.cs
@@ -1,13 +1,11 @@
 ﻿using System.Reflection;
 using System.Text;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Streetcode.BLL.Middleware
+namespace Streetcode.WebApi.Middleware.ApiRequestResponseMiddleware
 {
     public class ApiRequestResponseMiddleware
     {
@@ -49,9 +47,9 @@ namespace Streetcode.BLL.Middleware
             await next(context);
 
             var response = await GetResponseAsTextAsync(context.Response);
-            var filteredResponse = GetFilteredBody(response);
             await responseBody.CopyToAsync(originalBodyStream);
-            return filteredResponse;
+
+            return response;
         }
 
         private async Task<string> GetRequestAsTextAsync(HttpRequest request)
